// src/app/calculators/car-economy-calc/models/auxiliar.ts
export class City {
    description: string;
}

export class CityParamsPop {
    id?: number;
    TPH: number;
    PAX: number;
    travelDistance: number;
    travelTime: number;
    ASP: number;
    GNVPrize: number;
    alcoholPrize: number;
    gasolinePrize: number;
    cityID: string;
}

export class CityParamsUber {
    id?: number;
    TPH: number;
    PAX: number;
    travelDistance: number;
    travelTime: number;
    ASP: number;
    GNVPrize: number;
    alcoholPrize: number;
    gasolinePrize: number;
    cityID: string;
}